% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_grid_data}
\alias{extract_grid_data}
\title{Extract data for a single watershed}
\usage{
extract_grid_data(station, grid_data)
}
\arguments{
\item{station}{Name of a station (e.g. "1.48")}

\item{grid_data}{Data from one BIL file}
}
\value{
Vector of values for one watershed
}

