% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data.R
\name{get_metadata}
\alias{get_metadata}
\title{Load metadata for a set of stations}
\usage{
get_metadata(regine_main = meta_data$regine_main)
}
\arguments{
\item{Vector}{with stations (if not provided, the default returns all stations available)}
}
\value{
A dataframe with metadata
}
\examples{
regine_main <- c('1.48','1.49','1.50')
metadata <- get_metadata(regine_main)
}

