% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data.R
\name{load_data_mean}
\alias{load_data_mean}
\title{Load meteorological data for a set of stations}
\usage{
load_data_mean(path, station, time_vec)
}
\arguments{
\item{path}{Path to folder with BIL files}

\item{station}{Vector with stations}

\item{time_vec}{Vector with dates}
}
\value{
A list with data
}
\examples{
path = "//hdata/grid/metdata/met_obs_v2.0"
station = c("1.48","1.49","1.50")
time_vec = seq(as.Date("2011-01-01"), as.Date("2011-01-04"), by="days")
res = load_data_mean(path,station,time_vec)

}

