% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data.R
\name{load_flood_data}
\alias{load_flood_data}
\title{Load flood data for a set of stations}
\usage{
load_flood_data(regine_main)
}
\arguments{
\item{regine_main}{Vector with stations (regine_area and main_no seperated by a full stop)}

\item{path_met}{Path to folder with gridded meteorological data (requires netcdf files, and SeNorge version 2.0 and 2.1)}

\item{path_runoff}{Path to folder with files containing runoff data generated by lescon_var}

\item{time_vec}{Vector with dates (using format from package lubridate)}
}
\value{
A list with data for each station
}
\description{
Loads flood forecasting data for all available models
}
\examples{
regine_main <- c('1.48','1.49','1.50')
res <- load_flood_data(regine_main)
# Or to run for every station:
res <- load_flood_data(meta_data$regine_main)
# I need to check why, but many stations which should have been modelled have not been detected
# To see some which stations should have been modelled do:
which(meta_data$br9_Flomvarsling == "Y")
# Not many of them work. For exmaple 41 and 42 work so to see some modelling results you can do
res[[41]]$HBV$modelled
}

