% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_single_wsh}
\alias{load_single_wsh}
\title{Extract data for a single watershed from gridded meteorological data}
\usage{
load_single_wsh(regine_main, grid_data)
}
\arguments{
\item{regine_main}{Station name (regine_area and main_no seperated by a full stop)}

\item{grid_data}{Meteorological data stored as a vector (see functions read_bil_file and read_nc_file for loading such data)}
}
\value{
Vector of values
}

