% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_flomtabell}
\alias{read_flomtabell}
\title{Read Flomtabell.rap to get return levels for each station}
\usage{
read_flomtabell(filename = system.file("demodata", "flomtabell.txt", package =
  "NVEDATA"))
}
\arguments{
\item{filename}{Full path and name of the datafile (default is demodata/flomtabell.rap). This function gets data for all stations.
If needed, we could add a parameter to get only specific stations}
}
\value{
NOTHING Saves the results to an .rda file
}

