% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_HBV50_data}
\alias{read_HBV50_data}
\title{Read HBV50 modelling results}
\usage{
read_HBV50_data(path = "../Flood_forecasting/data/usikkerhet_grd/ut_test",
  filename = "vfpost_usikkerhet.txt")
}
\arguments{
\item{path}{Path to the HBV modelling results ('../Flood_forecasting/data/usikkerhet_grd/ut_test' as default)}

\item{filename}{Name of the file with HBV modelling results ("vfpost_usikkerhet.txt" as default))}
}
\value{
A dataframe with the modelling results and parameters
}

