% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flood_forecasting_data.R
\name{read_HBV_data}
\alias{read_HBV_data}
\title{Read HBV modelling results}
\usage{
read_HBV_data(filename = system.file("demodata/usikkerhet_grd/utskrift",
  "vfpost_usikkerhet.txt", package = "NVEDATA"))
}
\arguments{
\item{filename}{Full path and name of the file with HBV modelling results
Default is "demodata/usikkerhet_grd/utskrift/vfpost_usikkerhet.txt"}
}
\value{
HBV A dataframe with the modelling results, input and state parameters
}
\examples{
HBV <- read_HBV_data()
}

