% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_nc_file}
\alias{read_nc_file}
\title{Read netdcf file}
\usage{
read_nc_file(filename)
}
\arguments{
\item{filename}{Name of the netcdf file}
}
\value{
Vector with values (reshaped into a vector from a grid)
}

