% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flood_forecasting_data.R
\name{read_past_HBV}
\alias{read_past_HBV}
\title{Read HBV results of the past year for each station}
\usage{
read_past_HBV(folder = system.file("demodata/HBV_past_year", package =
  "NVEDATA"))
}
\arguments{
\item{folder}{Path to the folder where all the individual files are.
Default is "demodata/HBV_past_year"}
}
\value{
HBV_past_year A dataframe with HBV modelling results for the past year
Runoff_Obs is the observed discharge
Runoff_Sim.obs is the modelled discharge using observed meteorological data
Runoff_Sim.sim is the modelled discharge using raw forecast meteorological data
Runoff_Sim.sim.corr is the modelled discharge using corrected forecast meteorological data
}
\examples{
HBV_past_year <- read_past_HBV()
}

