% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_runoff_file}
\alias{read_runoff_file}
\title{Load runoff data from lescon_var output files}
\usage{
read_runoff_file(path, filename)
}
\arguments{
\item{path}{Path to folder with files}

\item{filename}{File name in format "regine_area.main_no.point_no.param_key.version_end_no"}

\item{time_vec}{Vector with times defining the period for data selection}
}
\value{
Data frame (dplyr) with time and runoff
}

