% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_data.R
\name{update_model_data}
\alias{update_model_data}
\title{Update data for models}
\usage{
update_model_data(path_met, path_runoff, path_model_data, overlap)
}
\arguments{
\item{path_met}{Path to folder with gridded meteorological data (requires netcdf files, and SeNorge version 2.0 and 2.1)}

\item{path_runoff}{Path to folder with files containing runoff data generated by lescon_var}

\item{path_model_data}{Path to folder in which to save the output files}

\item{overlap}{Number of days in the old data to replace with new data}
}
\value{
Data for runnig hydrological models
}
\examples{
path_met <- '//hdata/grid/metdata/met_obs_v2.0'
path_runoff <- '//hdata/fou/Avrenningskart/Data/Runoff_All'
path_model_data <- 'C:/Users/jmg/Desktop/testdata'   # Change to approprite folder
regine_main <- c('1.48','1.49','1.50')
time_vec <- seq(ymd(Sys.Date()-10), ymd(Sys.Date()-5), by = "days")
overlap <- 2
write_model_data(path_met, path_runoff, path_model_data, regine_main, time_vec)
update_model_data(path_met, path_runoff, path_model_data, overlap)
}

