% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data.R
\name{load_data_mean}
\alias{load_data_mean}
\title{Load data for a set of stations, average meteorological grids for the watershed}
\usage{
load_data_mean(path_met, path_runoff, regine_main, time_vec)
}
\arguments{
\item{path_met}{Path to folder with gridded meteorological data (requires netcdf files, and SeNorge version 2.0 and 2.1)}

\item{path_runoff}{Path to folder with files containing runoff data generated by lescon_var}

\item{regine_main}{Vector with stations (regine_area and main_no seperated by a full stop)}

\item{time_vec}{Vector with dates (using format from package lubridate)}
}
\value{
A list with data for each station
}
\description{
Runoff values stored on date yyyy-mm-dd are valid for the period from yyyy-mm-dd 00:00 to yyyy-mm-dd+1 00:00

Meteorological values stored on date yyyy-mm-dd are valid for the period from yyyy-mm-dd-1 06:00 to yyyy-mm-dd 06:00
}
\examples{
library(lubridate)
path_met <- '//hdata/grid/metdata/met_obs_v2.0'
path_runoff <- '//hdata/fou/Vannbalansekart/Data/Runoff_All'
regine_main <- c('1.48','1.49','1.50')
time_vec <- seq(ymd("2011-01-01"), ymd("2011-01-04"), by = "days")
res <- load_data_mean(path_met, path_runoff, regine_main, time_vec)
}

