% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flood_forecasting_data.R
\name{load_flood_data}
\alias{load_flood_data}
\title{Load flood data for a set of stations}
\usage{
load_flood_data(regine_main = meta_data$regine_main,
  HBV_2014_filename = "//hdata/drift/flom/usikkerhet_grd/utskrift/vfpost_usikkerhet.txt",
  HBV_2016_filename = "//hdata/drift/flom/usikkerhet_grd/ut_test/vfpost_usikkerhet.txt",
  HBV_P_filename = "//hdata/drift/flom/usikkerhet_grd/ut_test/vfp3030.txt",
  DDD_filename = "//hdata/drift/flom//DDD24h2015R/24hres.txt")
}
\arguments{
\item{regine_main}{Vector of station numbers (regine_area and main_no separated by a full stop: see example).
Default is meta_data$regine_main from the packaged meta_data}

\item{HBV_2014_filename}{File with HBV_2014 results
Default is the network location: '//hdata/drift/flom/usikkerhet_grd/utskrift/vfpost_usikkerhet.txt'}

\item{HBV_2016_filename}{File with HBV_2016 results
Default is the network location: '//hdata/drift/flom/usikkerhet_grd/ut_test/vfpost_usikkerhet.txt'}

\item{HBV_P_filename}{File with HBV_2016 precipitation correction results
Default is the network location: '//hdata/drift/flom/usikkerhet_grd/ut_test/vfp3030.txt'}

\item{DDD_filename}{File with DDD results
Default is the network location: '//hdata/drift/flom//DDD24h2015R/24hres.txt'}
}
\value{
NULL The function saves several .rda files to the working directory
}
\description{
Loads flood forecasting data for all available models
This function concatenates the results of all other functions defined in the file "read_flood_forecasting_data.R".
It is the only function from the NVEDATA package to be called directly by the flood forecasting app
}
\examples{
# For a specific set of stations:
regine_main <- c('1.48','1.49','1.50')
load_flood_data(regine_main)
# Or to run for every station:
load_flood_data()
# The following files should now be available in your working directory:
data_all.RData, meta_data.rda, HBV_2014.RData, HBV_2016.RData, HBV_past_year.RData, DDD.RData and flomtabell.RData
}

