% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_single_wsh}
\alias{load_single_wsh}
\title{Extract data for a single watershed}
\usage{
load_single_wsh(regine_main, grid_data)
}
\arguments{
\item{regine_main}{Name of a station given as 'regine_area.main_no' (e.g. '1.48')}

\item{grid_data}{Data from one BIL file stored as a vector}
}
\value{
Vector of values for one watershed
}

