% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_bil_file}
\alias{read_bil_file}
\title{Read BIL file}
\usage{
read_bil_file(filename, nrow = 1550, ncol = 1195)
}
\arguments{
\item{filename}{Name of the BIL file}

\item{nrow}{Number of rows in grid file (SeNorge grids used as default)}

\item{ncol}{Number of columns in grid file (SeNorge grids used as default)}
}
\value{
Vector with values (not reshaped into a grid)
}

