% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flood_forecasting_data.R
\name{read_DDD}
\alias{read_DDD}
\title{Read DDD modelling results}
\usage{
read_DDD(filename = system.file("demodata/DDD24h2015R", "24hres.txt", package
  = "NVEDATA"))
}
\arguments{
\item{filename}{Full path and name of the file with DDD modelling results
Defaults is "demodata/DDD24h2015R/24hres.txt"}
}
\value{
DDD A dataframe with modelling results, input and model state parameters
}
\examples{
DDD <- read_DDD()
}

