% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_flood_forecasting_data.R
\name{read_flomtabell}
\alias{read_flomtabell}
\title{Read Flomtabell.rap to get return levels for each station}
\usage{
read_flomtabell(filename = system.file("demodata", "flomtabell.txt", package =
  "NVEDATA"))
}
\arguments{
\item{filename}{Full path and name of the datafile
Default is "demodata/flomtabell.rap" which has data for all stations in Norway
If needed, we could add a parameter to get only specific stations}
}
\value{
flomtabell A dataframe with flood return levels
}
\examples{
flomtabell <- read_flomtabell()
}

