% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_HBV_data}
\alias{read_HBV_data}
\title{Read HBV modelling results}
\usage{
read_HBV_data(path = "./data/usikkerhet_grd/utskrift",
  filename = "vfpost_usikkerhet.txt")
}
\arguments{
\item{path}{Path to the HBV modelling results ('../Flood_forecasting/data/usikkerhet_grd/utskrift' as default)}

\item{filename}{Name of the file with HBV modelling results ("vfpost_usikkerhet.txt" as default))}
}
\value{
A dataframe with the modelling results and parameters
}

