% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_HBV_P}
\alias{read_HBV_P}
\title{Read HBV modelling results for -+ 50 percent precipitation}
\usage{
read_HBV_P(filename = system.file("demodata/usikkerhet_grd/ut_test",
  "vfp3030.txt", package = "NVEDATA"))
}
\arguments{
\item{filename}{Full path and name of the file with HBV_2016 +-50% modelling results ("demodata/usikkerhet_grd/ut_test/vfp3030.txt" as default))}
}
\value{
A dataframe with the modelling results and parameters
}

