% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_past_HBV}
\alias{read_past_HBV}
\title{Read Flomtabell.rap to get return levels for each station}
\usage{
read_past_HBV(folder = "./demodata/HBV_past_year")
}
\arguments{
\item{folder}{Path to the folder where all the individual files are. Default is demodata/HBV_past_year}
}
\value{
NOTHING Saves the results to an .rda file
}

